[e2e]reorg telemetry tests
