+++ tests/integration/telemetry/common/api/stats_wasm_filter_test.go
@@ -15,25 +15,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package nullvm
+package telemetryapi
 
 import (
 	"testing"
 
 	"istio.io/istio/pkg/test/framework"
 	"istio.io/istio/pkg/test/framework/components/istio"
-	"istio.io/istio/pkg/test/framework/features"
 	"istio.io/istio/pkg/test/framework/label"
 	"istio.io/istio/pkg/test/framework/resource"
-	common "istio.io/istio/tests/integration/telemetry/stats/prometheus"
+	"istio.io/istio/tests/integration/telemetry/common"
 )
 
-// TestStatsFilter verifies the stats filter could emit expected client and server side metrics.
+// TestTelemetryAPIStats verifies the stats filter could emit expected client and server side
+// metrics when configured with the Telemetry API (with EnvoyFilters disabled)
 // This test focuses on stats filter and metadata exchange filter could work coherently with
-// proxy bootstrap config. To avoid flake, it does not verify correctness of metrics, which
-// should be covered by integration test in proxy repo.
-func TestStatsFilter(t *testing.T) {
-	common.TestStatsFilter(t, features.Feature("observability.telemetry.stats.prometheus.http.nullvm"))
+// proxy bootstrap config with Wasm runtime. To avoid flake, it does not verify correctness
+// of metrics, which should be covered by integration test in proxy repo.
+func TestTelemetryAPIStats(t *testing.T) {
+	common.TestStatsFilter(t, "observability.telemetry.stats.common.http.nullvm")
 }
 
 func TestMain(m *testing.M) {
@@ -41,15 +41,29 @@ func TestMain(m *testing.M) {
 		Label(label.CustomSetup).
 		Label(label.IPv4). // https://github.com/istio/istio/issues/35915
 		Setup(istio.Setup(common.GetIstioInstance(), setupConfig)).
+		Setup(func(ctx resource.Context) error {
+			i, err := istio.Get(ctx)
+			if err != nil {
+				return err
+			}
+			return ctx.ConfigIstio().YAML(i.Settings().SystemNamespace, `
+apiVersion: telemetry.istio.io/v1alpha1
+kind: Telemetry
+metadata:
+  name: mesh-default
+spec:
+  metrics:
+  - providers:
+    - name: common
+`).Apply()
+		}).
 		Setup(common.TestSetup).
 		Run()
 }
 
-func setupConfig(_ resource.Context, cfg *istio.Config) {
+func setupConfig(c resource.Context, cfg *istio.Config) {
 	if cfg == nil {
 		return
 	}
-	// enable telemetry v2 with nullvm
-	cfg.Values["telemetry.v2.metadataExchange.wasmEnabled"] = "false"
-	cfg.Values["telemetry.v2.prometheus.wasmEnabled"] = "false"
+	cfg.Values["telemetry.v2.enabled"] = "false"
 }
