+++ tests/integration/telemetry/common/customizemetrics/testdata/attributegen_envoy_filter.yaml
