+++ tests/integration/telemetry/common/nullvm/istioctl_metrics_test.go
@@ -26,7 +26,7 @@ import (
 	"istio.io/istio/pkg/test/framework"
 	"istio.io/istio/pkg/test/framework/components/istioctl"
 	"istio.io/istio/pkg/test/util/retry"
-	common "istio.io/istio/tests/integration/telemetry/stats/prometheus"
+	"istio.io/istio/tests/integration/telemetry/common"
 )
 
 // TestIstioctlMetrics contains a basic validation of the experimental
