+++ tests/integration/telemetry/common/testdata/istio-mtls-dest-rule.yaml
