+++ tests/integration/telemetry/common/testdata/istio-mtls-gateway.yaml
