+++ tests/integration/telemetry/common/testdata/istio-mtls-vs.yaml
