+++ tests/integration/telemetry/common/wasm/stats_wasm_tcp_filter_test.go
@@ -15,18 +15,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package nullvm
+package wasm
 
 import (
 	"testing"
 
-	common "istio.io/istio/tests/integration/telemetry/stats/prometheus"
+	"istio.io/istio/tests/integration/telemetry/common"
 )
 
-func TestTcpMetric(t *testing.T) {
-	common.TestStatsTCPFilter(t, "observability.telemetry.stats.prometheus.tcp")
+func TestWASMTcpMetric(t *testing.T) { // nolint:interfacer
+	common.TestStatsTCPFilter(t, "observability.telemetry.stats.common.tcp")
 }
 
-func TestGatewayTCP(t *testing.T) {
-	common.TestStatsGatewayServerTCPFilter(t, "observability.telemetry.stats.prometheus.tcp")
+func TestWASMGatewayTCP(t *testing.T) {
+	common.TestStatsGatewayServerTCPFilter(t, "observability.telemetry.stats.common.tcp")
 }
