+++ tests/integration/telemetry/common/wasm/testdata/bad-filter.yaml
