+++ tests/integration/telemetry/common/wasm/testdata/registry-secret.yaml
