+++ tests/integration/telemetry/common/wasm/testdata/wasm-filter.yaml
