+++ tests/integration/telemetry/common/wasm/wasm_modules/header_injector/Dockerfile
