+++ tests/integration/telemetry/stackdriver/api/stackdriver_filter_test.go
@@ -34,8 +34,8 @@ import (
 	"istio.io/istio/pkg/test/framework/resource"
 	"istio.io/istio/pkg/test/scopes"
 	"istio.io/istio/pkg/test/util/retry"
-	"istio.io/istio/tests/integration/telemetry"
 	stackdrivertest "istio.io/istio/tests/integration/telemetry/stackdriver"
+	"istio.io/istio/tests/integration/telemetry/util"
 )
 
 const (
@@ -64,7 +64,7 @@ func TestStackdriverMonitoring(t *testing.T) {
 							return err
 						}
 						clName := cltInstance.Config().Cluster.Name()
-						trustDomain := telemetry.GetTrustDomain(cltInstance.Config().Cluster, stackdrivertest.Ist.Settings().SystemNamespace)
+						trustDomain := util.GetTrustDomain(cltInstance.Config().Cluster, stackdrivertest.Ist.Settings().SystemNamespace)
 						scopes.Framework.Infof("Validating for cluster %s", clName)
 
 						// Validate cluster names in telemetry below once https://github.com/istio/istio/issues/28125 is fixed.
